Removes concept of "default handler"
Because the `MiddlewarePipe` does not compose a "default handler", and
because the `MiddlewarePipe` is what the request handler runner
dispatches, there is no longer a concept of a "default delegate" (v2) or
"default handler" (v3); there is only middleware.

As such, this patch merges the `NotFoundHandler` into the `NotFoundMiddleware`.

A key is kept for the `DefaultDelegate` service, but it will now resolve
to the `NotFoundMiddleware`.

Fixes #441.
