+++ src/Container/NotFoundMiddlewareFactory.php
@@ -10,13 +10,28 @@ declare(strict_types=1);
 namespace Zend\Expressive\Container;
 
 use Psr\Container\ContainerInterface;
-use Zend\Expressive\Handler\NotFoundHandler;
+use Psr\Http\Message\ResponseInterface;
 use Zend\Expressive\Middleware\NotFoundMiddleware;
+use Zend\Expressive\Template\TemplateRendererInterface;
 
 class NotFoundMiddlewareFactory
 {
     public function __invoke(ContainerInterface $container) : NotFoundMiddleware
     {
-        return new NotFoundMiddleware($container->get(NotFoundHandler::class));
+        $config   = $container->has('config') ? $container->get('config') : [];
+        $renderer = $container->has(TemplateRendererInterface::class)
+            ? $container->get(TemplateRendererInterface::class)
+            : null;
+        $template = $config['zend-expressive']['error_handler']['template_404']
+            ?? NotFoundMiddleware::TEMPLATE_DEFAULT;
+        $layout = $config['zend-expressive']['error_handler']['layout']
+            ?? NotFoundMiddleware::LAYOUT_DEFAULT;
+
+        return new NotFoundMiddleware(
+            $container->get(ResponseInterface::class),
+            $renderer,
+            $template,
+            $layout
+        );
     }
 }
