--- src/Handler/NotFoundHandler.php
@@ -1,100 +1,0 @@
-<?php
-/**
- * @see       https://github.com/zendframework/zend-expressive for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive/blob/master/LICENSE.md New BSD License
- */
-
-declare(strict_types=1);
-
-namespace Zend\Expressive\Handler;
-
-use Fig\Http\Message\StatusCodeInterface;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
-use Psr\Http\Server\RequestHandlerInterface;
-use Zend\Expressive\Template\TemplateRendererInterface;
-
-class NotFoundHandler implements RequestHandlerInterface
-{
-    public const TEMPLATE_DEFAULT = 'error::404';
-    public const LAYOUT_DEFAULT = 'layout::default';
-
-    /**
-     * @var TemplateRendererInterface
-     */
-    private $renderer;
-
-    /**
-     * This duplicates the property in StratigilityNotFoundHandler, but is done
-     * to ensure that we have access to the value in the methods we override.
-     *
-     * @var ResponseInterface
-     */
-    protected $responsePrototype;
-
-    /**
-     * @var string
-     */
-    private $template;
-
-    /**
-     * @var string
-     */
-    private $layout;
-
-    public function __construct(
-        ResponseInterface $responsePrototype,
-        TemplateRendererInterface $renderer = null,
-        string $template = self::TEMPLATE_DEFAULT,
-        string $layout = self::LAYOUT_DEFAULT
-    ) {
-        $this->responsePrototype = $responsePrototype;
-        $this->renderer = $renderer;
-        $this->template = $template;
-        $this->layout = $layout;
-    }
-
-    /**
-     * Creates and returns a 404 response.
-     */
-    public function handle(ServerRequestInterface $request) : ResponseInterface
-    {
-        if (! $this->renderer) {
-            return $this->generatePlainTextResponse($request);
-        }
-
-        return $this->generateTemplatedResponse($request);
-    }
-
-    /**
-     * Generates a plain text response indicating the request method and URI.
-     */
-    private function generatePlainTextResponse(ServerRequestInterface $request) : ResponseInterface
-    {
-        $response = $this->responsePrototype->withStatus(StatusCodeInterface::STATUS_NOT_FOUND);
-        $response->getBody()
-            ->write(sprintf(
-                'Cannot %s %s',
-                $request->getMethod(),
-                (string) $request->getUri()
-            ));
-
-        return $response;
-    }
-
-    /**
-     * Generates a response using a template.
-     *
-     * Template will receive the current request via the "request" variable.
-     */
-    private function generateTemplatedResponse(ServerRequestInterface $request) : ResponseInterface
-    {
-        $response = $this->responsePrototype->withStatus(StatusCodeInterface::STATUS_NOT_FOUND);
-        $response->getBody()->write(
-            $this->renderer->render($this->template, ['request' => $request, 'layout' => $this->layout])
-        );
-
-        return $response;
-    }
-}
